chore!: update tree-sitter to latest
BREAKING CHANGE: Users that use the `tree-sitter` feature may have to
refactor their code based on the breaking changes between `0.24.3` and
`0.25.3`.

Users that bring in `tree-sitter` as a seperate dependency may get
compile errors related to the version mismatch. The used `tree-sitter`
version should match the version provided by `texter` which is `0.25.3`
in order to use the exposed methods.